pyproject.toml
[tool.poetry]
name = "music-analysis-agent"
version = "0.1.0"
description = ""
authors = ["Johannes Naylor <johannes@tapped.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.0.345"
openai = "^1.3.7"
python-dotenv = "^1.0.0"
google-search-results = "^2.4.2"
numexpr = "^2.8.7"
langchainhub = "^0.1.14"
pydantic = "^2.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
